// src/app/shared/models/meeting.model.ts
export class Meeting {
    id: number;
    organizer_id: number;
    title: string;
    description: string;
    start_date: Date;
    end_date: Date;
    location: string;
    members: any;
    avatar: string;
    firstname: string;
    lastname: string;
    client_id: number;
    status: number;
}